+++ src/main/types.ts
@@ -5,6 +5,7 @@ import { VideoCategory } from '../types/VideoCategory';
 import { Tag } from 'react-tag-autocomplete';
 import { DateValueType } from 'react-tailwindcss-datepicker';
 import { ESupportedEncoders } from './obsEnums';
+import { ReactNode } from 'react';
 
 /**
  * Application recording status.
@@ -479,7 +480,7 @@ type VideoMarker = {
 
 type SliderMark = {
   value: number;
-  label: JSX.Element;
+  label: ReactNode;
 };
 
 type CloudStatus = {
@@ -493,6 +494,7 @@ type CloudStatus = {
   del: boolean;
   usage: number;
   limit: number;
+  migrated: boolean;
 };
 
 type DiskStatus = {